requirements.txt
Pillow
PyYAML
Requests
addict
aiohttp
chardet
dynamicprompts
ftfy
ipdb
numpy
opencv_contrib_python
opencv_python
opencv_python_headless
pandas
pefile
piexif
pipreqs
pycocotools
pyexiv2
pytorch_lightning
pytz
regex
segment_anything
setuptools
supervision
timm
tomli
torch
torchvision
tqdm
transformers
ultralytics
yapf
openai-clip
huggingface_hub
gguf>=0.9.1
imagesize
sentencepiece
accelerate
cpm-kernels
diffusers
bitsandbytes
peft>=0.10.0
pyrallis
mmengine
open_clip_torch
omegaconf
fal-client
anthropic
openai
google-genai>=1.65.0
dotenv
elevenlabs
python-magic-bin
deepface